pyproject.toml
[tool.poetry]
name = "reportgen"
version = "0.1.0"
description = "A library for generating PDF reports from JSON configurations"
authors = ["Ron Reiter"]
readme = "README.md"
packages = [{include = "reportgen"}]

[tool.poetry.dependencies]
python = ">=3.9"
weasyprint = "^63.1"
altair = "^5.0.0"
vl-convert-python = "^1.7.0"
pandas = "^2.0.0"
pydantic = "^2.0.0"
jinja2 = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
isort = "^5.0.0"

[tool.poetry.scripts]
reportgen = "reportgen.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
